product views refactoring

+++ shop_app/products/views.py
@@ -1,31 +1,30 @@
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse
-
 from .models import Product
 from .forms import ProductForm
 
 def product_list(request):
-    products = Product.objects.all()  # Pobieramy wszystkie produkty
+    """Lista produktów"""
+    products = Product.objects.all()
     return render(request, 'products/product_list.html', {'products': products})
 
-
 def product_detail(request, pk):
-    product = get_object_or_404(Product, pk=pk)  # Pobieramy produkt na podstawie jego ID
+    """Szczegóły produktu"""
+    product = get_object_or_404(Product, pk=pk)
     return render(request, 'products/product_detail.html', {'product': product})
 
-
 def product_add(request):
+    """Dodawanie nowego produktu"""
     if request.method == 'POST':
         form = ProductForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect('product_list')  # Przekierowanie na stronę listy produktów
+            return redirect('product_list')
     else:
         form = ProductForm()
     return render(request, 'products/product_add.html', {'form': form})
 
-
 def product_edit(request, pk):
+    """Edycja produktu"""
     product = get_object_or_404(Product, pk=pk)
     if request.method == "POST":
         form = ProductForm(request.POST, instance=product)
@@ -36,8 +35,10 @@ def product_edit(request, pk):
         form = ProductForm(instance=product)
     return render(request, 'products/product_edit.html', {'form': form, 'product': product})
 
-
 def product_delete(request, pk):
+    """Usuwanie produktu"""
     product = get_object_or_404(Product, pk=pk)
-    product.delete()
-    return redirect('product_list')  # Przekierowanie na stronę listy produktów
+    if request.method == "POST":
+        product.delete()
+        return redirect('product_list')
+    return render(request, 'products/product_confirm_delete.html', {'product': product})
